+++ packages/common/src/templates/_internal/computeLinks.ts
@@ -25,17 +25,20 @@ export function computeLinks(
     token = session.token;
   }
 
+  // If a solution template is deployed, we don't support
+  // managing it in the workspace, so we kick users to AGO
+  const isDeployed = item.typeKeywords?.includes("Deployed");
+  const itemHomeUrl = getItemHomeUrl(item.id, requestOptions);
   return {
-    self: getItemHomeUrl(item.id, requestOptions),
+    self: itemHomeUrl,
     siteRelative: getHubRelativeUrl(
       item.type,
       getItemIdentifier(item),
       item.typeKeywords
     ),
-    workspaceRelative: getRelativeWorkspaceUrl(
-      item.type,
-      getItemIdentifier(item)
-    ),
+    workspaceRelative: isDeployed
+      ? itemHomeUrl
+      : getRelativeWorkspaceUrl(item.type, getItemIdentifier(item)),
     thumbnail: getItemThumbnailUrl(item, requestOptions, token),
   };
 }