+++ packages/common/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @esri/hub-common [14.32.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.32.0...@esri/hub-common@14.32.1) (2023-10-10)
+
+
+### Bug Fixes
+
+* deployed solution redirects ([#1268](https://github.com/Esri/hub.js/issues/1268)) ([8df3607](https://github.com/Esri/hub.js/commit/8df3607b764a2dbfc743f6d0ece03b634e8924d1))
+
+# @esri/hub-common [14.32.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.31.0...@esri/hub-common@14.32.0) (2023-10-09)
+
+
+### Features
+
+* support for Site and Hub user app resources ([#1260](https://github.com/Esri/hub.js/issues/1260)) ([4a314e1](https://github.com/Esri/hub.js/commit/4a314e1d62d46374988a49cb1eb202d62a95cdb4))
+
 # @esri/hub-common [14.31.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.30.1...@esri/hub-common@14.31.0) (2023-10-09)